+++ src/Indice.Features.Cases.UI/CasesUIOptions.cs
@@ -13,14 +13,14 @@ public class CasesUIOptions : SpaUIOptions
 
     /// <summary>A list containing the tags of the canvases that should be displayed in the dashboard.</summary>
     /// <remarks>If left null or empty all canvases are displayed.</remarks>
-    public List<string> Canvases { get; set; }
+    public List<string> DashboardTags { get; set; }
 
     /// <summary>Creates a new instance <see cref="CasesUIOptions"/>.</summary>
     public CasesUIOptions() {
         ConfigureIndexParameters = args => {
             args[$"%({nameof(ApiUrl)})"] = ApiUrl;
             args[$"%({nameof(Lang)})"] = Lang;
-            args[$"%({nameof(Canvases)})"] = Canvases?.Count is 0 ? null : string.Join(',', Canvases);
+            args[$"%({nameof(DashboardTags)})"] = DashboardTags?.Count is 0 ? null : string.Join(',', DashboardTags);
         };
     }
 }